requirements.txt
requests==2.31.0
python-dotenv==1.0.0
geopy==2.4.1
pyodbc==5.0.1 